pyproject.toml
[tool.poetry]
name = "demoauth"
version = "0.1.0"
description = ""
authors = ["dev-ravan <velmurugan1211r@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pymongo = {version = "3.12", extras = ["srv"]}
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "<4.0.0"
certifi = "^2024.8.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
